requirements.txt
# https://cloud.google.com/python/docs/reference/logging/latest
google-cloud-logging>=3.11.1

# https://github.com/cloudevents/sdk-python
cloudevents==1.11.0

# https://pypi.org/project/flask
Flask==3.0.2

# https://pypi.org/project/gunicorn
gunicorn==21.2.0